# agentauth/docs/openapi.yaml
openapi: 3.1.0
info:
  title: AgentAuth API
  version: 0.7.0
  description: |
    **AgentAuth** is a production-ready authentication and identity service for AI agents.

    ## Features
    - JWT-based authentication
    - Rate limiting (100 req/15min general, 10 req/15min auth)
    - Webhooks for events
    - Activity logging
    - Permission-based access control
    - **Persona System** — define and manage an agent's digital soul (v0.7.0)
    - **ZKP Anonymous Verification** — prove identity without revealing credentials (v0.7.0)
    - **Anti-Drift Vault** — detect behavioral drift and auto-revoke rogue agents (v0.7.0)

    ## Authentication
    Most endpoints require an API key in the `X-Api-Key` header:
    ```
    X-Api-Key: ag_sk_your_api_key_here
    ```

    ## Rate Limits
    - **General endpoints**: 100 requests per 15 minutes
    - **Authentication endpoints**: 10 requests per 15 minutes

    Rate limit info is included in response headers:
    - `X-RateLimit-Limit`: Max requests allowed
    - `X-RateLimit-Remaining`: Requests remaining
    - `X-RateLimit-Reset`: Time until limit resets (Unix timestamp)

    ## Base URL
    - **Production**: `https://api.agentauth.dev`
    - **Local**: `http://localhost:3000`

  contact:
    name: AgentAuth Support
    email: umytbaynazarow754@gmail.com
    url: https://agentauth.dev
  license:
    name: MIT
    url: https://github.com/umyt-dev/Agent-Identity/blob/main/LICENSE
servers:
  - url: https://api.agentauth.dev
...
paths:
...
  /v1/agents/register:
    post:
      summary: Register a new agent
      description: |
        Register a new AI agent and receive an API key for authentication.

        **Important:** The API key is only shown once. Save it securely!

        **Rate limit:** 10 requests per 15 minutes
      operationId: registerAgent
      tags:
        - Agents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRegistrationRequest'
            example:
              name: CustomerSupportAgent
              owner_email: admin@company.com
              permissions:
                - read:messages
                - write:responses
              tier: free
      responses:
        '201':
          description: Agent registered successfully
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRegistrationResponse'
              example:
                agent_id: ag_1a2b3c4d5e6f
                name: CustomerSupportAgent
                owner_email: admin@company.com
                api_key: ag_sk_1a2b3c4d5e6f7g8h9i0j
                tier: free
                permissions:
                  - read:messages
                  - write:responses
                created_at: '2026-02-01T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/agents/verify:
    post:
      summary: Verify agent credentials
      description: |
        Verify agent credentials (agent_id + API key) and receive agent details.

        **Rate limit:** 10 requests per 15 minutes
      operationId: verifyAgent
      tags:
        - Agents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentVerificationRequest'
            example:
              agent_id: ag_1a2b3c4d5e6f
              api_key: ag_sk_1a2b3c4d5e6f7g8h9i0j
      responses:
        '200':
          description: Agent verified successfully
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentVerificationResponse'
              example:
                valid: true
                agent_id: ag_1a2b3c4d5e6f
                name: CustomerSupportAgent
                tier: free
                permissions:
                  - read:messages
                  - write:responses
                status: active
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/agents:
    get:
      summary: List all agents
      description: |
        List all registered agents with pagination.

        **Note:** This endpoint may require admin authentication in future versions.
      operationId: listAgents
      tags:
        - Agents
      parameters:
        - name: limit
          in: query
          description: Maximum number of agents to return
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of agents to skip (for pagination)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: status
          in: query
          description: Filter by agent status
          required: false
          schema:
            type: string
            enum: [active, inactive, suspended]
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
              example:
                agents:
                  - agent_id: ag_1a2b3c4d5e6f
                    name: CustomerSupportAgent
                    owner_email: admin@company.com
                    tier: free
                    status: active
                    created_at: '2026-02-01T12:00:00.000Z'
                total: 1
                limit: 50
                offset: 0
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/agents/{agent_id}:
    get:
      summary: Get agent details
      description: Retrieve detailed information about a specific agent
      operationId: getAgent
      tags:
        - Agents
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete agent
      description: Permanently delete an agent
      operationId: deleteAgent
      tags:
        - Agents
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
              example:
                success: true
                message: Agent deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/agents/{agent_id}/tier:
    put:
      summary: Update agent tier
      description: Update the pricing tier for an agent
      operationId: updateAgentTier
      tags:
        - Agents
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tier
              properties:
                tier:
                  type: string
                  enum: [free, pro, enterprise]
            example:
              tier: pro
      responses:
        '200':
          description: Tier updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent_id:
                    type: string
                  tier:
                    type: string
                  updated_at:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/agents/{agent_id}/status:
    put:
      summary: Update agent status
      description: Activate, deactivate, or suspend an agent
      operationId: updateAgentStatus
      tags:
        - Agents
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [active, inactive, suspended]
            example:
              status: suspended
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent_id:
                    type: string
                  status:
                    type: string
                  updated_at:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/agents/{agent_id}/permissions:
    put:
      summary: Update agent permissions
      description: Update the permission scopes for an agent
      operationId: updateAgentPermissions
      tags:
        - Agents
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - permissions
              properties:
                permissions:
                  type: array
                  items:
                    type: string
                  example:
                    - read:messages
                    - write:responses
                    - delete:messages
      responses:
        '200':
          description: Permissions updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent_id:
                    type: string
                  permissions:
                    type: array
                    items:
                      type: string
                  updated_at:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/agents/{agent_id}/activity:
    get:
      summary: Get agent activity logs
      description: Retrieve authentication and activity logs for an agent
      operationId: getAgentActivity
      tags:
        - Agents
      parameters:
        - $ref: '#/components/parameters/AgentId'
        - name: limit
          in: query
          description: Maximum number of logs to return
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          description: Number of logs to skip
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Activity logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActivityLog'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ─── Persona (Soul Layer) ────────────────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Agent API key
  parameters:
    AgentId:
      name: agent_id
      in: path
      description: Unique agent identifier
      required: true
      schema:
        type: string
        pattern: '^ag_[a-zA-Z0-9]+$'
      example: ag_1a2b3c4d5e6f
...
  headers:
    X-RateLimit-Limit:
      description: Maximum number of requests allowed in the current window
      schema:
        type: integer
      example: 10
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current window
      schema:
        type: integer
      example: 7
    X-RateLimit-Reset:
      description: Unix timestamp when the rate limit resets
      schema:
        type: integer
      example: 1738329600
  schemas:
...
    AgentRegistrationRequest:
      type: object
      required: [name, owner_email]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Agent name
        owner_email:
          type: string
          format: email
          description: Owner's email address
        permissions:
          type: array
          items:
            type: string
          description: Permission scopes
          default: []
        tier:
          type: string
          enum: [free, pro, enterprise]
          default: free
    AgentRegistrationResponse:
      type: object
      required: [agent_id, name, owner_email, api_key, tier, permissions, created_at]
      properties:
        agent_id:
          type: string
        name:
          type: string
        owner_email:
          type: string
          format: email
        api_key:
          type: string
          description: API key (only shown once)
        tier:
          type: string
          enum: [free, pro, enterprise]
        permissions:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    AgentVerificationRequest:
      type: object
      required: [agent_id, api_key]
      properties:
        agent_id:
          type: string
        api_key:
          type: string
    AgentVerificationResponse:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        agent_id:
          type: string
        name:
          type: string
        tier:
          type: string
          enum: [free, pro, enterprise]
        permissions:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, suspended]
        error:
          type: string
    Agent:
      type: object
      required: [agent_id, name, owner_email, tier, status, created_at]
      properties:
        agent_id:
          type: string
        name:
          type: string
        owner_email:
          type: string
          format: email
        tier:
          type: string
          enum: [free, pro, enterprise]
        permissions:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, suspended, revoked]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgentListResponse:
      type: object
      required: [agents, total, limit, offset]
      properties:
        agents:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    ActivityLog:
      type: object
      required: [id, agent_id, success, timestamp]
      properties:
        id:
          type: string
        agent_id:
          type: string
        success:
          type: boolean
        reason:
          type: string
        ip_address:
          type: string
        timestamp:
          type: string
          format: date-time

    # ─── Persona Schemas ─────────────────────────────────────────────────────
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
  responses:
    BadRequest:
      description: Bad request - validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Validation failed
            message: Invalid input data
            details:
              - field: email
                message: Must be a valid email address
    Unauthorized:
      description: Unauthorized - invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Invalid agent credentials
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not Found
            message: Resource not found
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          schema:
            type: integer
            example: 0
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Rate Limit Exceeded
            message: Too many requests. Please try again later.
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal Server Error
            message: An unexpected error occurred